requirements.txt
django
django-ipgeolocation
django-ratelimit
celery
redis
requests
